+++ src/plugins/lokijs/lokijs-helper.ts
@@ -29,12 +29,13 @@ export const LOKI_KEY_OBJECT_BROADCAST_CHANNEL_MESSAGE_TYPE = 'rxdb-lokijs-remot
  * Loki attaches a $loki property to all data
  * which must be removed before returning the data back to RxDB.
  */
-export function stripLokiKey<T>(docData: T & { $loki?: number }): T {
+export function stripLokiKey<T>(docData: T & { $loki?: number; $lastWriteAt?: number; }): T {
     if (!docData.$loki) {
         return docData;
     }
     const cloned = flatClone(docData);
     delete cloned.$loki;
+    delete cloned.$lastWriteAt;
     return cloned;
 }