+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Bugfixes:
   - Compound primary key migration throws "Value of primary key(s) cannot be changed" [#3546](https://github.com/pubkey/rxdb/pull/3546) Thanks [@nothingkid](https://github.com/nothingkid)
   - Allow `_id` as primaryKey [#3562](https://github.com/pubkey/rxdb/pull/3562) Thanks [@SuperKirik](https://github.com/SuperKirik)
 
+Other:
+  - LokiJS: All documents are stored with a `$lastWriteAt` field, so we can implement an auto compaction later.
+
 ### 10.5.4 (30 November 2021)
 
 Bugfixes: